+++ packages/core/src/internal/validation/futures/validateArtifactContractAt.ts
