+++ packages/core/src/internal/validation/futures/validateNamedContractAt.ts
