+++ packages/core/src/internal/validation/futures/validateNamedContractCall.ts
