+++ packages/core/src/internal/validation/futures/validateNamedStaticCall.ts
