+++ packages/core/src/internal/validation/futures/validateReadEventArgument.ts
