+++ packages/core/src/internal/validation/futures/validateSendData.ts
